--- src/Microsoft.Docs.Template/Models/GitUserInfo.cs
@@ -1,16 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.Docs.Build
-{
-    public class GitUserInfo
-    {
-        public string Name { get; set; }
-
-        public string ProfileUrl { get; set; }
-
-        public string DisplayName { get; set; }
-
-        public string Id { get; set; }
-    }
-}
