+++ src/Microsoft.Docs.Template/Models/PageModel.cs
@@ -44,4 +44,15 @@ namespace Microsoft.Docs.Build
 
         public JObject Metadata { get; set; }
     }
+
+    public class GitUserInfo
+    {
+        public string Name { get; set; }
+
+        public string ProfileUrl { get; set; }
+
+        public string DisplayName { get; set; }
+
+        public string Id { get; set; }
+    }
 }
